requirements.txt
# search_agent
strands-agents
strands-agents-tools
strands-agents[a2a]
strands-agents-tools[a2a_client]
ddgs
streamlit
nest-asyncio